refactor: improve project migration feature architecture
 - Split project migration into separate components:
   - MigrationV2Modal for migration projects in Renku 2.0
   - MigrationV1Modal for migration projects in Renku Legacy
- Improve state management:
  - Add useMigrationForm hook for form logic
  - Move logic for Gitlab projects to GitlabProjectList component
